% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.r
\name{join_regex}
\alias{join_regex}
\title{Join regexes together with an "or"}
\usage{
join_regex(...)
}
\arguments{
\item{...}{regex strings to join together}
}
\description{
Simply concatenates regex strings with a "|" symbol.
}
