% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.r
\name{tokenize_twitter}
\alias{tokenize_twitter}
\title{Tokenize list of tweets}
\usage{
tokenize_twitter(x, stopwords = NULL, stem_language = "porter")
}
\arguments{
\item{x}{list of documents}

\item{stopwords}{vector of words to remove}

\item{stem_language}{which language to use (if any) for stemming using \code{\link[SnowballC]{wordStem}}}
}
\description{
Tokenize list of tweets
}
